% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{copy_file}
\alias{copy_file}
\title{Copy file}
\usage{
copy_file(access_key_id, secret_access_key, src_bucket_name, src_key_name,
  dst_bucket_name, dst_key_name, is_ordinary_calling_format = FALSE,
  region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{src_bucket_name}{Source S3 bucket name}

\item{src_key_name}{Source S3 key name}

\item{dst_bucket_name}{Destination S3 bucket name}

\item{dst_key_name}{Destination S3 key name}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\value{
a list of content copy information
}
\description{
\code{copy_file} copies a file within S3.
}
\details{
The source and destination bucket can be the same. A file can be copied into a subfolder of the destination bucket by adjusting \code{dst_key_name}. eg) \code{dst_key_name = 'subfolder/dst_key_name'}.

For \code{is_ordinary_calling_format} and \code{region}, see \link{connect_test}.
}
\examples{
\dontrun{

copy_file('access-key-id', 'secret-access-key', 'src-bucket-name', 'src-key-name', 'dst-bucket-name', 'dst-key-name')
copy_file('access-key-id', 'secret-access-key', 'same-bucket-name', 'src-key-name', 'same-bucket-name', 'subfolder/src-key-name')
}
}

