% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{create_bucket}
\alias{create_bucket}
\title{Create S3 bucket}
\usage{
create_bucket(access_key_id, secret_access_key, bucket_name, location = NULL,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{bucket_name}{S3 bucket name}

\item{location}{S3 location}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\value{
a list of bucket creation information
}
\description{
\code{create_bucket} creates a S3 bucket.
}
\details{
A bucket can be created in a different location by specifying \code{location} - see \link{get_location}.

For \code{is_ordinary_calling_format} and \code{region}, see \link{connect_test}.
}
\examples{
\dontrun{

create_bucket('aws-access-id', 'secret-access-key', 'bucket-name')
create_bucket('aws-access-id', 'secret-access-key', 'bucket-name', 'location')
}
}

