% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{delete_key}
\alias{delete_key}
\title{Delete key}
\usage{
delete_key(access_key_id, secret_access_key, bucket_name, key_name = NULL,
  prefix = NULL, is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{delete_key}
}
\examples{
\dontrun{

}
}

