% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{delete_keys}
\alias{delete_keys}
\title{Delete S3 key}
\usage{
delete_keys(access_key_id, secret_access_key, bucket_name, key_name = NULL,
  prefix = NULL, is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{bucket_name}{S3 bucket name}

\item{key_name}{S3 key name}

\item{prefix}{prefix that filters keys}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\value{
a list of key deletion information
}
\description{
\code{delete_keys} deletes a key or keys.
}
\details{
If \code{key_name} is provided, only the key is deleted. If \code{prefix} is provided, all keys that are found by the given prefix are deleted. Finally, if \code{prefix} is \code{NULL}, all keys in the bucket are deleted.

For \code{is_ordinary_calling_format} and \code{region}, see \link{connect_test}.
}
\examples{
\dontrun{

delete_keys('aws-access-id', 'secret-access-key', 'bucket-name', key_name = 'key-name')
delete_keys('aws-access-id', 'secret-access-key', 'bucket-name', prefix = 'prefix')
delete_keys('aws-access-id', 'secret-access-key', 'bucket-name')
}
}

