% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{download_file}
\alias{download_file}
\title{Download file}
\usage{
download_file(access_key_id, secret_access_key, bucket_name, key_name = NULL,
  file_path = NULL, pattern = NULL, prefix = NULL,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{download_file}
}
\examples{
\dontrun{

}
}

