% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{download_files}
\alias{download_files}
\title{Download S3 contents}
\usage{
download_files(access_key_id, secret_access_key, bucket_name, key_name = NULL,
  file_path = NULL, pattern = NULL, prefix = NULL,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{bucket_name}{S3 bucket name}

\item{key_name}{S3 key name}

\item{file_path}{download file path}

\item{pattern}{key search pattern}

\item{prefix}{prefix that filters keys}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\value{
a data frame of content download information
}
\description{
\code{download_file} download one or more S3 contents.
}
\details{
If \code{key_name} is provided, only the key is downloaded. If \code{prefix} is provided, all keys that are found by the given prefix are downloaded.
If \code{pattern} is provided, only the keys that match the pattern are downloaded. Finally, if only \code{bucket_name} is provided, all keys in the bucket are downloaded.

For \code{is_ordinary_calling_format} and \code{region}, see \link{connect_test}.
}
\examples{
\dontrun{

download_file('aws-access-key', 'secret-access-key', bucket_name = 'bucket-name', key_name = 'key-name', file_path = set_file_path(getwd()))
download_file('aws-access-key', 'secret-access-key', bucket_name = 'bucket-name', prfix = 'prefix', file_path = set_file_path(getwd()))
download_file('aws-access-key', 'secret-access-key', bucket_name = 'bucket-name', prfix = 'prefix', pattern = 'pattern', file_path = set_file_path(getwd()))
download_file('aws-access-key', 'secret-access-key', bucket_name = 'bucket-name', file_path = set_file_path(getwd()))
}
}

