% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{get_dir_str}
\alias{get_dir_str}
\title{Get directory string}
\usage{
get_dir_str(dir_str = getwd())
}
\arguments{
\item{dir_str}{directory string}
}
\value{
directory string where slash or backslash is corrected
}
\description{
\code{get_dir_str} converts a slash into a backslash (or vice versa) according to the operating system. This function is not exported.
}

