% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{lookup_key}
\alias{lookup_key}
\title{Lookup S3 key}
\usage{
lookup_key(access_key_id, secret_access_key, bucket_name, key_name,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{bucket_name}{S3 bucket name}

\item{key_name}{S3 key name}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\value{
a list of reponse including status
}
\description{
\code{lookup_key} checks whether a key exists in a bucket.
}
\details{
For \code{is_ordinary_calling_format} and \code{region}, see \link{connect_test}.
}
\examples{
\dontrun{

lookup_key('aws-access-id', 'secret-access-key', 'bucket-name', 'key-name')
}
}

