% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{lookup_location}
\alias{lookup_location}
\title{Lookup location}
\usage{
lookup_location()
}
\value{
a charactor vector of available locations
}
\description{
\code{lookup_location} returns available locations where a bucket can be created.
}

