% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{lookup_region}
\alias{lookup_region}
\title{Lookup region}
\usage{
lookup_region()
}
\value{
a character vector of available regions
}
\description{
\code{lookup_region} returns available regions where connection to AWS can be made. See \link{connect_test}.
}

