% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs3helper.R
\name{rs3helper}
\alias{rs3helper}
\title{Python wrapper of boto s3 library}
\description{
\code{rs3helper} is a collection of functions on AWS S3.
Each function is backed by a Python script that performs an action on the boto library - boto is probably the most popular interface to Amazon Web Services in Python.
In this regards, this package can be considered as a Python port of boto S3 library.
For those who are interested in native R, please see the aws.s3 package of the cloudyr project.
}
\details{
Further details on boto and aws.s3 can be found below.

An Introduction to boto S3 - \url{http://boto.cloudhackers.com/en/latest/s3_tut.html}

AWS S3 Client Package - \url{https://github.com/cloudyr/aws.s3}
}

