% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{rs3wrapper}
\alias{rs3wrapper}
\title{Wrapper of other functions}
\usage{
rs3wrapper(access_key_id, secret_access_key,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\value{
a wrapper functional of other functions
}
\description{
\code{rs3wrapper} returns a functional with connection related arguments pre-filled.
}
\details{
The following 4 connection related variables are repeated in almost all functions and their values can be applied at once using this functions.
A partially applied functional is returned where the 4 values are pre-filled by executing \code{rs3wrapper}. Then a function and its argument can be entered in each of the two braces.

For \code{is_ordinary_calling_format} and \code{region}, see \link{connect_test}.
}
\examples{
\dontrun{

wrapper <- rs3wrapper('access-key-id', 'secret-access-key')
wrapper(get_all_buckets)()
wrapper(lookup_bucket)(bucket_name = 'bucket-name')
}
}

