% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{set_access_control_list}
\alias{set_access_control_list}
\title{Set access control list}
\usage{
set_access_control_list(access_key_id, secret_access_key, bucket_name,
  permission, key_name = NULL, is_ordinary_calling_format = FALSE,
  region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{set_access_control_list}
}
\examples{
\dontrun{

}
}

