% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{set_file_path}
\alias{set_file_path}
\title{Set a file path}
\usage{
set_file_path(file_path = getwd(), file_name = NULL)
}
\arguments{
\item{file_path}{file path}

\item{file_name}{file name}
}
\value{
file path with the correct separator of the OS
}
\description{
\code{set_file_path} constructs a file path with the correct separator (slash or backslash) of the OS.
}
\examples{
\dontrun{

set_file_path(file_path = getwd(), file_name = NULL)
}
}

