% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{upload_file}
\alias{upload_file}
\title{Upload file}
\usage{
upload_file(access_key_id, secret_access_key, bucket_name, file_path, file_name,
  prefix = NULL, is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{upload_file}
}
\examples{
\dontrun{

}
}

