% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{connect_test}
\alias{connect_test}
\title{S3 connection test}
\usage{
connect_test(access_key_id, secret_access_key,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\value{
connection string
}
\description{
\code{connect_test} is a function to test connection to AWS.
}
\details{
AWS key pairs are necessary. For accessing a bucket whose name is non-DNS compliant, set \code{is_ordinary_calling_format} to be \code{TRUE}.
For example, if a bucket name contains a dot(.) or includes both small and capital letters, it is not DNS compliant. For details, see \link{http://docs.aws.amazon.com/AmazonS3/latest/dev/BucketRestrictions.html}).
Connection can also be made in a specific region. Check available regions by running \link{lookup_region}.
}
\examples{
\dontrun{

connect_test('access-key-id', 'secret-access-key')
}
}

