% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{copy_file}
\alias{copy_file}
\title{Copy file}
\usage{
copy_file(access_key_id, secret_access_key, src_bucket_name, src_key_name,
  dst_bucket_name, dst_key_name, is_ordinary_calling_format = FALSE,
  region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{copy_file}
}
\examples{
\dontrun{

}
}

