% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{create_bucket}
\alias{create_bucket}
\title{Create bucket}
\usage{
create_bucket(access_key_id, secret_access_key, bucket_name, location = NULL,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{create_bucket}
}
\examples{
\dontrun{

}
}

