% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{delete_bucket}
\alias{delete_bucket}
\title{Delete bucket}
\usage{
delete_bucket(access_key_id, secret_access_key, bucket_name,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{delete_bucket}
}
\examples{
\dontrun{

}
}

