% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{generate_url}
\alias{generate_url}
\title{Generate URL}
\usage{
generate_url(access_key_id, secret_access_key, bucket_name, key_name, seconds,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{generate_url}
}
\examples{
\dontrun{

}
}

