% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{get_access_control_list}
\alias{get_access_control_list}
\title{Get access control list}
\usage{
get_access_control_list(access_key_id, secret_access_key, bucket_name,
  key_name = NULL, is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{get_access_control_list}
}
\examples{
\dontrun{

}
}

