% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{get_all_buckets}
\alias{get_all_buckets}
\title{Get all buckets}
\usage{
get_all_buckets(access_key_id, secret_access_key,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{get_all_buckets}
}
\examples{
\dontrun{

}
}

