% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{get_keys}
\alias{get_keys}
\title{Get keys}
\usage{
get_keys(access_key_id, secret_access_key, bucket_name, prefix = NULL,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{get_keys}
}
\examples{
\dontrun{

}
}

