% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{lookup_bucket}
\alias{lookup_bucket}
\title{Lookup bucket}
\usage{
lookup_bucket(access_key_id, secret_access_key, bucket_name,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{bucket_name}{S3 bucket name}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\description{
\code{lookup_bucket} checks whether the given bucket exists. If the bucket name is non-DNS compliant, set \code{is_ordinary_calling_format} to be \code{TRUE}.
See \link{connect_test} for further details.
}
\examples{
\dontrun{

}
}

