% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{lookup_key}
\alias{lookup_key}
\title{Lookup key}
\usage{
lookup_key(access_key_id, secret_access_key, bucket_name, key_name,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{param}{param}
}
\description{
\code{lookup_key}
}
\examples{
\dontrun{

}
}

