% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3utils.R
\name{rs3wrapper}
\alias{rs3wrapper}
\title{Wrapper of other functions}
\usage{
rs3wrapper(access_key_id, secret_access_key,
  is_ordinary_calling_format = FALSE, region = NULL)
}
\arguments{
\item{access_key_id}{AWS access key id}

\item{secret_access_key}{AWS secret access key}

\item{is_ordinary_calling_format}{Connection calling format}

\item{region}{Connection region}
}
\value{
a wrapper function (\code{function(f, ...)})
}
\description{
\code{rs3wrapper} returns a function where connection related arguments are pre-filled.
}
\details{
The following 4 connection related arguments are required in all functions:\code{access_key_id}, \code{secret_access_key}, \code{is_ordinary_calling_format} and \code{region}.
The arguments can be pre-filled at once using this function. A function that has a function argument(\code{f}) and unspecified arguments (\code{...}) is returned.
Then it is possible to execute a function with the remaining arguments.

For \code{is_ordinary_calling_format} and \code{region}, see \link{connect_test}.
}
\examples{
\dontrun{

wrapper <- rs3wrapper('access-key-id', 'secret-access-key')
wrapper(get_all_buckets) # no remaining argument
wrapper(lookup_bucket, bucket_name = 'bucket-name') # bucket_name is specified
wrapper(lookup_key, 'bucket-name', 'key-name') # bucket_name and key_name are specified in order
}
}

